Adding ability to load books into new db layout Adding ability to download epubs from new db layout Adding associated UI pages for the new db layout Adding proxy-file caching for 3rd party files to avoid CORS related issues when fetching images Adding associated API endpoints for new db layout

+++ src/CardboardBox.Epub/EpubBuilder.Metadata.cs
@@ -24,7 +24,7 @@
 
 		IEpubBuilder Metadata(XElement element);
 
-		IEpubBuilder BelongsTo(string title, int position, string type = MetaData.COLLECTION_TYPE_SET);
+		IEpubBuilder BelongsTo(string title, long position, string type = MetaData.COLLECTION_TYPE_SET);
 	}
 
 	public partial class EpubBuilder
@@ -96,7 +96,7 @@
 			return this;
 		}
 
-		public IEpubBuilder BelongsTo(string title, int position, string type = MetaData.COLLECTION_TYPE_SET)
+		public IEpubBuilder BelongsTo(string title, long position, string type = MetaData.COLLECTION_TYPE_SET)
 		{
 			Content.Meta.CollectionTitle = title;
 			Content.Meta.CollectionType = type;
